pyproject.toml
[project]
name = "buildingapplicationswithaiagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "av>=16.0.1",
    "fastapi>=0.122.0",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.3",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "streamlit-webrtc>=0.64.4",
    "tiktoken>=0.12.0",
    "typing-extensions>=4.15.0",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]
